refactor(intelligence) Improve model name caching
