+++ intelligence/ts/src/engines/caching/storage.ts
@@ -21,55 +21,7 @@ export interface CachedEntry {
 export interface CachedMapping {
   mapping: Record<string, CachedEntry>;
 }
-
-export class CacheStorage {
-  protected async load(): Promise<CachedMapping | null> {
-    await Promise.resolve();
-    throw new Error('Method not implemented');
-  }
-  protected async save(_cache: CachedMapping): Promise<void> {
-    await Promise.resolve();
-    throw new Error('Method not implemented');
-  }
-
-  /**
-   * Gets a model mapping if it was saved in the cache, otherwise null.
-   */
-  async getItem(key: string): Promise<CachedEntry | null> {
-    let cache = await this.load();
-    if (!cache) {
-      cache = { mapping: {} };
-    }
-
-    if (key in cache.mapping) {
-      return cache.mapping[key];
-    }
-    return null;
-  }
-
-  /**
-   * Adds or updates a model mapping in the cache with the current timestamp.
-   */
-  async setItem(key: string, value: string): Promise<void> {
-    const now = Date.now();
-    let cache = await this.load();
-    if (!cache) {
-      cache = { mapping: {} };
-    }
-    cache.mapping[key] = { engineModel: value, timestamp: now };
-    await this.save(cache);
-  }
-
-  /**
-   * Removes a model from the cache.
-   */
-  async remove(model: string, engine: string): Promise<void> {
-    const cache = await this.load();
-    const key = `${model}_${engine}`;
-    if (cache && key in cache.mapping) {
-      const { [key]: removed, ...rest } = cache.mapping;
-      cache.mapping = rest;
-      await this.save(cache);
-    }
-  }
+export abstract class CacheStorage {
+  abstract getItem(key: string): Promise<CachedEntry | null>;
+  abstract setItem(key: string, value: string | null): Promise<void>;
 }
